[arcane,std] Déplace les classes 'BasicGenericReader' et 'BasicReader' dans le fichier 'BasicReader.cc'.
